deps/v8/test/inspector/debugger/pause-on-instrumentation-expected.txt

Test if pauses break execution when requested during instrumentation pause in js

Running test: testPauseDuringInstrumentationPause
Set instrumentation breakpoints and requested script evaluation.
Paused at foo.js with reason "instrumentation".
Requested debugger pause.
Resumed.
Paused at foo.js with reason "other".
Resumed.
Done.

Running test: testInstrumentationRemoveDuringInstrumentationPause
Paused at  with reason "instrumentation".
Removed instrumentation breakpoint
Resumed
Evaluation result: 42
